Split news parsing and praise helpers into separate modules
