+++ commands.c
@@ -4868,7 +4868,7 @@ cmd_upload(char *args)
 
     if (access(FILE_DB, W_OK | X_OK) == -1) {
         dlog(3, "cmd_upload: FILE_DB not writable/searchable: %s", FILE_DB);
-        output("\nKunde inte skriva till filkatalogen. Meddela SysOp.\n\n");
+        output("\n"MSG_FILE_DB_ER"\n\n");
 
         signal(SIGNAL_NEW_TEXT, baffo);
         signal(SIGNAL_NEW_MSG, newmsg);
@@ -4880,7 +4880,7 @@ cmd_upload(char *args)
      */
     if (UPLOADPRGM == NULL || *UPLOADPRGM == '\0') {
         dlog(3, "cmd_upload: UPLOADPRGM is empty");
-        output("\nUpload är inte konfigurerat på detta system.\n\n");
+        output("\n"MSG_FILES_OFF"\n\n");
 
         signal(SIGNAL_NEW_TEXT, baffo);
         signal(SIGNAL_NEW_MSG, newmsg);
@@ -4953,7 +4953,7 @@ cmd_upload(char *args)
             dlog(7, "cmd_upload: child exit=%d", WEXITSTATUS(status));
 
             if (WEXITSTATUS(status) == 127) {
-                output("\nUploadprogrammet kunde inte startas.\n\n");
+                output("\n"MSG_ULPRGMERROR"\n\n");
 
                 if (chdir(cwd) == -1) {
                     dlog(3, "cmd_upload: chdir back to cwd failed after exec failure: %s", cwd);
@@ -5848,7 +5848,7 @@ cmd_nethack(char *args)
 
     /* If still not found */
     if (access(nethack_path, X_OK) != 0) {
-        output("\nFel: Kan inte starta Nethack - har SysOp glömt att installera den?\n");
+        output("\n"MSG_NETHACKER01"\n");
         return 0;
     }
 
@@ -7245,13 +7245,13 @@ cmd_like(char *args)
     snprintf(confxtra, sizeof(confxtra), "%s/%d%s", SKLAFF_DB, Current_conf, CONFXTRA_FILE);
 
     if ((fd = open_file(confxtra, OPEN_CREATE)) == -1) {
-        output("\nCould not open confxtra, this is bad, contact your SysOp!\n\n");
+        output("\n"MSG_OPEN_C_E"\n\n");
         return 0;
     }
 
     if ((buf = read_file(fd)) == NULL) {
         close_file(fd);
-        output("\nCould not read confxtra, this is bad, contact your SysOp!\n\n");
+        output("\n"MSG_READ_C_E"\n\n");
         return 0;
     }
 
@@ -7283,7 +7283,7 @@ cmd_like(char *args)
         // Append new like to existing ![hiss] block
         FILE *fp = fopen(confxtra, "a");
         if (!fp) {
-            output("\nCould not write to confxtra, please let your Sysop know!\n\n");
+            output("\n"MSG_OPEN_C_E"\n\n");
             free(oldbuf);
             close_file(fd);
             return 0;
@@ -7295,7 +7295,7 @@ cmd_like(char *args)
         // No ![hiss] block — create it and append entry
         FILE *fp = fopen(confxtra, "a");
         if (!fp) {
-            output("\nCould not write to confxtra, please let your Sysop know!\n\n");
+            output("\n"MSG_READ_C_E"\n\n");
             free(oldbuf);
             close_file(fd);
             return 0;
